chore: decouple extension filters from versioning extension filters
